# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "cobbs.ai RESTful API",
        "description": "Swagger Docs of cobbs.ai Backend API",
        "version": "1.70.0",
        "contact": {
            "email": "baron@cobbs.ai"
        }
    },
    "basePath": "/api",
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/upload-file": {
            "post": {
                "tags": [
                    "File API"
                ],
                "description": "upload file to casdoor storage",
                "operationId": "ApiController.UploadFile",
                "parameters": [
                    {
                        "in": "formData",
                        "name": "file",
                        "description": "The base64 encoded file data",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "type",
                        "description": "The file type/extension",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "name",
                        "description": "The file name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Response object",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Response": {
            "title": "Response",
            "type": "object",
            "properties": {
                "data": {
                    "additionalProperties": {
                        "description": "support string, struct or []struct",
                        "type": "string"
                    }
                },
                "data2": {
                    "additionalProperties": {
                        "description": "support string, struct or []struct",
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}